[Feature-3134]Refactor Curator Zookeeper Client to be singleton
